requirements.txt
wheel==0.45.1
twine==5.1.1
PyYAML==6.0.2
python-dotenv==0.21.0
pandas